예외처리 추가
