+++ src/main/java/org/bbagisix/challenge/controller/ChallengeController.java
@@ -13,6 +13,7 @@ public class ChallengeController {
 
 	private final ChallengeService challengeService;
 
+	// ✅ 단일 챌린지 상세 조회 (User ID 없이 테스트용)
 	@GetMapping("/{challengeId}")
 	public ResponseEntity<ChallengeDTO> getChallengeById(@PathVariable Long challengeId) {
 		ChallengeDTO result = challengeService.getChallengeById(challengeId);
