+++ src/main/java/org/bbagisix/challenge/service/ChallengeService.java
@@ -1,9 +1,11 @@
 package org.bbagisix.challenge.service;
 
 import lombok.RequiredArgsConstructor;
+import org.bbagisix.challenge.domain.ChallengeVO;
 import org.bbagisix.challenge.dto.ChallengeDTO;
 import org.bbagisix.challenge.mapper.ChallengeMapper;
-import org.bbagisix.challenge.domain.ChallengeVO;
+import org.bbagisix.exception.BusinessException;
+import org.bbagisix.exception.ErrorCode;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -13,7 +15,15 @@ public class ChallengeService {
 	private final ChallengeMapper challengeMapper;
 
 	public ChallengeDTO getChallengeById(Long challengeId) {
+		if (challengeId == null) {
+			throw new BusinessException(ErrorCode.CHALLENGE_ID_REQUIRED);
+		}
+
 		ChallengeVO vo = challengeMapper.findByChallengeId(challengeId);
+		if (vo == null) {
+			throw new BusinessException(ErrorCode.CHALLENGE_NOT_FOUND);
+		}
+
 		return ChallengeDTO.from(vo);
 	}
 }
