+++ src/main/java/org/bbagisix/chat/service/ChatService.java
@@ -7,8 +7,8 @@ import org.bbagisix.chat.converter.ChatMessageConverter;
 import org.bbagisix.chat.domain.ChatMessageVO;
 import org.bbagisix.chat.dto.ChatMessageDTO;
 import org.bbagisix.chat.entity.ChatMessage;
-import org.bbagisix.chat.exception.BusinessException;
-import org.bbagisix.chat.exception.ErrorCode;
+import org.bbagisix.exception.BusinessException;
+import org.bbagisix.exception.ErrorCode;
 import org.bbagisix.chat.mapper.ChatMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
