+++ src/main/java/org/bbagisix/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package org.bbagisix.chat.exception;
+package org.bbagisix.exception;
 
 import java.time.LocalDateTime;
 
@@ -8,11 +8,13 @@ import lombok.Getter;
 @Getter
 @Builder
 public class ErrorResponse {
+
 	private final String code;
 	private final String message;
 	private final LocalDateTime timestamp;
-	private final String path;
+	private final String path; // Web API 경로 (null 허용)
 
+	// ✅ 일반 HTTP 에러 응답
 	public static ErrorResponse of(ErrorCode errorCode) {
 		return ErrorResponse.builder()
 			.code(errorCode.getCode())
@@ -30,14 +32,14 @@ public class ErrorResponse {
 			.build();
 	}
 
-	// WebSocket 용 에러 응답
+	// ✅ WebSocket 전용 에러 응답
 	@Getter
 	@Builder
 	public static class WebSocketErrorResponse {
-		private final String type;        // "ERROR"
-		private final String code;        // "C001"
-		private final String message;    // 사용자에게 보여줄 메시지
-		private final Long challengeId;    // 에러가 발생한 챌린지
+		private final String type;             // 고정값: "ERROR"
+		private final String code;             // 예: "CH001"
+		private final String message;          // 사용자에게 보여줄 메시지
+		private final Long challengeId;        // 관련 챌린지 ID
 		private final LocalDateTime timestamp;
 
 		public static WebSocketErrorResponse of(ErrorCode errorCode, Long challengeId) {
@@ -59,7 +61,5 @@ public class ErrorResponse {
 				.timestamp(LocalDateTime.now())
 				.build();
 		}
-
 	}
-
 }
