+++ src/main/resources/mappers/challenge/ChallengeMapper.xml
@@ -5,6 +5,7 @@
 
 <mapper namespace="org.bbagisix.challenge.mapper.ChallengeMapper">
 
+    <!-- 1. 챌린지 ID로 단일 챌린지 조회 -->
     <select id="findByChallengeId" parameterType="long" resultType="org.bbagisix.challenge.domain.ChallengeVO">
         SELECT
             challenge_id,
@@ -14,6 +15,16 @@
             description
         FROM challenge
         WHERE challenge_id = #{challengeId}
+    </select>
+
+    <!-- 2. 유저가 해당 챌린지에 참여 중인지 확인 -->
+    <select id="existsUserChallenge" resultType="boolean">
+        SELECT EXISTS (
+            SELECT 1
+            FROM user_challenge
+            WHERE challenge_id = #{challengeId}
+              AND user_id = #{userId}
+        )
     </select>
 
 </mapper>
